adminpage/pyproject.toml
[tool.poetry]
name = "adminpage"
version = "0.1.0"
description = ""
authors = ["one_zero_eight"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
django = "5.1.3"
django-admin-autocomplete-filter = "^0.7.1"
django-admin-rangefilter = "^0.13.2"
django-auth-adfs = "^1.14.0"
django-cors-headers = "^4.7.0"
django-hijack = "^3.7.0"
django-image-optimizer = "^1.0.3"
django-import-export = "^4.3.1"
django-pglock = "^1.7.1"
django-prometheus = "^2.3.1"
django-revproxy = "^0.13.0"
django-sendfile2 = "^0.7.1"
django-smartfields = { git = "https://github.com/ViniciusBernardo/django-smartfields.git", branch = "master" }
django-stubs = "^5.1.3"
django-tinymce = "^4.1.0"
djangorestframework-stubs = "^3.15.3"
drf-spectacular = {extras = ["sidecar"], version = "^0.28.0"}
openpyxl = "^3.1.5"
pillow = "9.5.0"
psycopg2-binary = "^2.9.10"
pyjwt = "2.4.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-freezegun = { git = "https://github.com/oizik/pytest-freezegun.git", branch = "master" }
pytz = "^2025.1"
requests = "^2.32.3"
sentry-sdk = "^2.19.0"
typing-extensions = "^4.12.2"
uvicorn = "^0.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
